ramdisk:cortex-init: adapt to cluster_plug

+++ WORKING-DIR/ramdisk/ramdisk/sbin/ext/cortexbrain-tune.sh
@@ -82,7 +82,9 @@ if [ "$(cat /data/gabriel_cortex_sleep)" -eq "1" ]; then
 
 	echo "$(cat /cache/fsync_enabled)" > /sys/module/sync/parameters/fsync_enabled;
 
-	echo "$(cat /cache/lc_corectl_state)" > /sys/devices/system/cpu/cpu4/core_ctl/max_cpus;
+	if [ -e /sys/devices/system/cpu/cpu4/core_ctl/max_cpus ];then
+		echo "$(cat /cache/lc_corectl_state)" > /sys/devices/system/cpu/cpu4/core_ctl/max_cpus;
+	fi;
 
 	echo "$(cat /cache/sched_epa)" > /proc/sys/kernel/sched_enable_power_aware;
 
@@ -105,8 +107,10 @@ SLEEP_MODE()
 	echo "$(cat /sys/module/sync/parameters/fsync_enabled)" > /cache/fsync_enabled;
 	echo "1" > /sys/module/sync/parameters/fsync_enabled;
 
-	echo "$(cat /sys/devices/system/cpu/cpu4/core_ctl/max_cpus)" > /cache/lc_corectl_state;
-	echo "0" > /sys/devices/system/cpu/cpu4/core_ctl/max_cpus;
+	if [ -e /sys/devices/system/cpu/cpu4/core_ctl/max_cpus ];then
+		echo "$(cat /sys/devices/system/cpu/cpu4/core_ctl/max_cpus)" > /cache/lc_corectl_state;
+		echo "0" > /sys/devices/system/cpu/cpu4/core_ctl/max_cpus;
+	fi;
 
 	echo "$(cat /proc/sys/kernel/sched_enable_power_aware)" > /cache/sched_epa;
 	echo "1" > /proc/sys/kernel/sched_enable_power_aware;
